+++ cppfand/Graph.h
@@ -55,7 +55,7 @@ struct GraphD
 };
 
 struct ViewPortType {
-	integer x1 = 0, y1 = 0, x2 = 0, y2 = 0;
+	short x1 = 0, y1 = 0, x2 = 0, y2 = 0;
 	bool Clip = false;
 };
