+++ cppfand/oaccess.h
@@ -30,7 +30,7 @@ void SetIsSQLFile();
 #endif
 WORD GetCatIRec(pstring Name, bool MultiLevel);
 WORD Generation();
-void TurnCat(WORD Frst, WORD N, integer I);
+void TurnCat(WORD Frst, WORD N, short I);
 std::string RdCatField(WORD CatIRec, FieldDescr* CatF);
 void WrCatField(FileD* catFD, WORD CatIRec, FieldDescr* CatF, const std::string& Txt);
 void WrCatField(WORD CatIRec, FieldDescr* CatF, std::string Txt);
